requirements.txt
enum34
python-snappy
msgpack-python
